% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_nn_resample.R
\name{private_nn_resample}
\alias{private_nn_resample}
\title{Resample points to preserve privacy}
\usage{
private_nn_resample(x, id = NULL, K = 5, size = nrow(x))
}
\arguments{
\item{x}{a two-column matrix of coordinates}

\item{id}{a vector of integers representing group membership (default NULL)}

\item{K}{the number of nearest neighbours each kernel must overlap with
default 5.}

\item{size}{the number of points to sample. Default \code{nrow(x)}}
}
\value{
a two-column matrix of resampled points
}
\description{
This function takes a set of x,y coordinates and resamples them
according to a nearest-neighbour kernel, so that points in sparser
parts of the distribution have a larger bandwidth, and their
kernels overlap with high probability on the kernels of at least
K other points
}
